moved base MO arguments into utils
